+++ src/network/mod.rs
@@ -18,6 +18,7 @@ mod async_executor_strategy;
 mod cluster_connection;
 mod connection;
 mod connection_state;
+mod message_queue;
 mod network_handler;
 mod pub_sub_push;
 mod reconnection_state;